+++ BookSearchApp_UIKitTCA/BookSearchApp_UIKitTCA/App/SceneDelegate.swift
@@ -1,0 +1,40 @@
+//
+//  SceneDelegate.swift
+//  BookSearchApp_UIKitTCA
+//
+//  Created by 송형욱 on 2022/11/14.
+//
+
+import UIKit
+
+class SceneDelegate: UIResponder, UIWindowSceneDelegate {
+
+    var window: UIWindow?
+
+
+    func scene(_ scene: UIScene, willConnectTo session: UISceneSession, options connectionOptions: UIScene.ConnectionOptions) {
+        
+        guard let _ = (scene as? UIWindowScene) else { return }
+    }
+
+    func sceneDidDisconnect(_ scene: UIScene) {
+        
+    }
+
+    func sceneDidBecomeActive(_ scene: UIScene) {
+        
+    }
+
+    func sceneWillResignActive(_ scene: UIScene) {
+       
+    }
+
+    func sceneWillEnterForeground(_ scene: UIScene) {
+        
+    }
+
+    func sceneDidEnterBackground(_ scene: UIScene) {
+        
+    }
+}
+
