+++ BookSearchApp_UIKitTCA/BookSearchApp_UIKitTCA/Resources/Assets.xcassets/AppIcon.appiconset/Contents.json
